+++ shim.js
@@ -20,7 +20,7 @@ if (typeof process === 'undefined') {
   global.process = require('process');
 } else {
   const bProcess = require('process');
-  for (const p in bProcess) {
+  for (var p in bProcess) {
     if (!(p in process)) {
       process[p] = bProcess[p];
     }
@@ -29,7 +29,6 @@ if (typeof process === 'undefined') {
 
 process.browser = false;
 if (typeof Buffer === 'undefined') global.Buffer = require('buffer').Buffer;
-
 // global.location = global.location || { port: 80 }
 const isDev = typeof __DEV__ === 'boolean' && __DEV__;
 process.env.NODE_ENV = isDev ? 'development' : 'production';