+++ src/model/keychain.ts
@@ -1,0 +1,165 @@
+import { captureException, captureMessage } from '@sentry/react-native';
+import { forEach, isNil } from 'lodash';
+import {
+  getAllInternetCredentials,
+  getAllInternetCredentialsKeys,
+  getInternetCredentials,
+  hasInternetCredentials,
+  Options,
+  resetInternetCredentials,
+  setInternetCredentials,
+  UserCredentials,
+} from 'react-native-keychain';
+import { delay } from '../helpers/utilities';
+import logger from 'logger';
+
+interface AnonymousKey {
+  length: number;
+  nil: boolean;
+  type: string;
+}
+
+interface AnonymousKeyData {
+  [key: string]: AnonymousKey;
+}
+
+export async function saveString(
+  key: string,
+  value: string,
+  accessControlOptions: Options
+): Promise<void> {
+  return new Promise(async (resolve, reject) => {
+    try {
+      await setInternetCredentials(key, key, value, accessControlOptions);
+      logger.sentry(`Keychain: saved string for key: ${key}`);
+      resolve();
+    } catch (e) {
+      logger.sentry(`Keychain: failed to save string for key: ${key}`, e);
+      captureMessage('Keychain write first attempt failed');
+      await delay(1000);
+      try {
+        await setInternetCredentials(key, key, value, accessControlOptions);
+        logger.sentry(
+          `Keychain: saved string for key: ${key} on second attempt`
+        );
+        resolve();
+      } catch (e) {
+        logger.sentry(`Keychain: failed to save string for key: ${key}`, e);
+        captureMessage('Keychain write second attempt failed');
+        reject(e);
+      }
+    }
+  });
+}
+
+export async function loadString(
+  key: string,
+  options?: Options
+): Promise<null | string> {
+  try {
+    const credentials = await getInternetCredentials(key, options);
+    if (credentials) {
+      logger.log(`Keychain: loaded string for key: ${key}`);
+      return credentials.password;
+    }
+    logger.sentry(`Keychain: string does not exist for key: ${key}`);
+  } catch (err) {
+    logger.sentry(
+      `Keychain: failed to load string for key: ${key} error: ${err}`
+    );
+    captureException(err);
+  }
+  return null;
+}
+
+export async function saveObject(
+  key: string,
+  value: Object,
+  accessControlOptions: Options
+): Promise<void> {
+  const jsonValue = JSON.stringify(value);
+  return saveString(key, jsonValue, accessControlOptions);
+}
+
+export async function loadObject(
+  key: string,
+  options?: Options
+): Promise<null | Object> {
+  const jsonValue = await loadString(key, options);
+  if (!jsonValue) return null;
+  try {
+    const objectValue = JSON.parse(jsonValue);
+    logger.log(`Keychain: parsed object for key: ${key}`);
+    return objectValue;
+  } catch (err) {
+    logger.sentry(
+      `Keychain: failed to parse object for key: ${key} error: ${err}`
+    );
+    captureException(err);
+  }
+  return null;
+}
+
+export async function remove(key: string): Promise<void> {
+  try {
+    await resetInternetCredentials(key);
+    logger.log(`Keychain: removed value for key: ${key}`);
+  } catch (err) {
+    logger.log(
+      `Keychain: failed to remove value for key: ${key} error: ${err}`
+    );
+    captureException(err);
+  }
+}
+
+async function loadAllKeys(): Promise<null | UserCredentials[]> {
+  try {
+    const response = await getAllInternetCredentials();
+    if (response) {
+      return response.results;
+    }
+  } catch (err) {
+    logger.sentry(`Keychain: failed to loadAllKeys error: ${err}`);
+    captureException(err);
+  }
+  return null;
+}
+
+export async function getAllKeysAnonymized(): Promise<null | AnonymousKeyData> {
+  const data: AnonymousKeyData = {};
+  const results = await loadAllKeys();
+  forEach(results, result => {
+    data[result?.username] = {
+      length: result?.password?.length,
+      nil: isNil(result?.password),
+      type: typeof result?.password,
+    };
+  });
+  return data;
+}
+
+export async function loadAllKeysOnly(): Promise<null | string[]> {
+  try {
+    const response = await getAllInternetCredentialsKeys();
+    if (response) {
+      return response.results;
+    }
+  } catch (err) {
+    logger.log(`Keychain: failed to loadAllKeys error: ${err}`);
+    captureException(err);
+  }
+  return null;
+}
+
+export async function hasKey(key: string) {
+  try {
+    const result = await hasInternetCredentials(key);
+    return result;
+  } catch (err) {
+    logger.sentry(
+      `Keychain: failed to check if key ${key} exists -  error: ${err}`
+    );
+    captureException(err);
+  }
+  return null;
+}