+++ src/redux/walletconnect.js
@@ -18,6 +18,7 @@ import {
   saveWalletConnectSession,
 } from '../handlers/localstorage/walletconnectSessions';
 import { sendRpcCall } from '../handlers/web3';
+import { convertDappNameToDisplay } from '../helpers/dappNameHandler';
 import WalletTypes from '../helpers/walletTypes';
 import { getFCMToken } from '../model/firebase';
 import { Navigation } from '../navigation';
@@ -107,34 +108,32 @@ export const walletConnectOnSessionRequest = (
         if (error) throw error;
         const { peerId, peerMeta } = payload.params[0];
         const imageUrl = get(peerMeta, 'icons[0]');
-
-        InteractionManager.runAfterInteractions(() => {
-          Navigation.handleAction(Routes.WALLET_CONNECT_APPROVAL_SHEET, {
-            callback: async approved => {
-              if (approved) {
-                dispatch(setPendingRequest(peerId, walletConnector));
-                dispatch(walletConnectApproveSession(peerId, callback));
-                analytics.track('Approved new WalletConnect session', {
-                  dappName: peerMeta.name,
-                  dappUrl: peerMeta.url,
-                });
-              } else {
-                await dispatch(
-                  walletConnectRejectSession(peerId, walletConnector)
-                );
-                callback && callback('reject');
-                analytics.track('Rejected new WalletConnect session', {
-                  dappName: peerMeta.name,
-                  dappUrl: peerMeta.url,
-                });
-              }
-            },
-            meta: {
-              dappName: peerMeta.name,
-              dappUrl: peerMeta.url,
-              imageUrl,
-            },
-          });
+        const name = convertDappNameToDisplay(peerMeta.name);
+        Navigation.handleAction(Routes.WALLET_CONNECT_APPROVAL_SHEET, {
+          callback: async approved => {
+            if (approved) {
+              dispatch(setPendingRequest(peerId, walletConnector));
+              dispatch(walletConnectApproveSession(peerId, callback));
+              analytics.track('Approved new WalletConnect session', {
+                dappName: name,
+                dappUrl: peerMeta.url,
+              });
+            } else {
+              await dispatch(
+                walletConnectRejectSession(peerId, walletConnector)
+              );
+              callback && callback('reject');
+              analytics.track('Rejected new WalletConnect session', {
+                dappName: name,
+                dappUrl: peerMeta.url,
+              });
+            }
+          },
+          meta: {
+            dappName: name,
+            dappUrl: peerMeta.url,
+            imageUrl,
+          },
         });
       });
     } catch (error) {