+++ src/screens/ChangeWalletSheet.js
@@ -1,3 +1,4 @@
+import { captureException } from '@sentry/react-native';
 import { get } from 'lodash';
 import React, { useCallback, useMemo, useRef, useState } from 'react';
 import { InteractionManager } from 'react-native';
@@ -12,6 +13,7 @@ import { Modal } from '../components/modal';
 import { Sheet, SheetTitle } from '../components/sheet';
 import { Text } from '../components/text';
 import { removeWalletData } from '../handlers/localstorage/removeWallet';
+import showWalletErrorAlert from '../helpers/support';
 import WalletTypes from '../helpers/walletTypes';
 import { useWalletsWithBalancesAndNames } from '../hooks/useWalletsWithBalancesAndNames';
 import { createWallet } from '../model/wallet';
@@ -332,17 +334,28 @@ export default function ChangeWalletSheet() {
                     return false;
                   });
 
-                // If we found it, use it to create the new account
-                if (primaryWalletKey) {
-                  await dispatch(
-                    createAccountForWallet(primaryWalletKey, color, name)
-                  );
-                  await initializeWallet();
-                  // If doesn't exist, we need to create a new wallet
-                } else {
-                  await createWallet(null, color, name);
-                  await dispatch(walletsLoadState());
-                  await initializeWallet();
+                try {
+                  // If we found it, use it to create the new account
+                  if (primaryWalletKey) {
+                    await dispatch(
+                      createAccountForWallet(primaryWalletKey, color, name)
+                    );
+                    await initializeWallet();
+                    // If doesn't exist, we need to create a new wallet
+                  } else {
+                    await createWallet(null, color, name);
+                    await dispatch(walletsLoadState());
+                    await initializeWallet();
+                  }
+                } catch (e) {
+                  dispatch(isCreatingAccount(false));
+                  logger.sentry('Error while trying to add account');
+                  captureException(e);
+                  if (selectedWallet.damaged) {
+                    setTimeout(() => {
+                      showWalletErrorAlert();
+                    }, 1000);
+                  }
                 }
               }
               creatingWallet.current = false;
@@ -365,6 +378,7 @@ export default function ChangeWalletSheet() {
     goBack,
     initializeWallet,
     navigate,
+    selectedWallet.damaged,
     selectedWallet.id,
     selectedWallet.primary,
     wallets,