+++ src/screens/ImportSeedPhraseSheet.js
@@ -7,7 +7,7 @@ import React, {
   useRef,
   useState,
 } from 'react';
-import { Alert, InteractionManager, StatusBar } from 'react-native';
+import { Alert, StatusBar } from 'react-native';
 import { KeyboardArea } from 'react-native-keyboard-area';
 import styled from 'styled-components/primitives';
 import ActivityIndicator from '../components/ActivityIndicator';
@@ -248,23 +248,32 @@ export default function ImportSeedPhraseSheet() {
         const input = resolvedAddress ? resolvedAddress : seedPhrase.trim();
         initializeWallet(input, color, name ? name : '')
           .then(success => {
+            handleSetImporting(false);
             if (success) {
-              analytics.track('Imported seed phrase', {
-                hadPreviousAddressWithValue,
-              });
-              InteractionManager.runAfterInteractions(() => {
+              setTimeout(() => {
                 navigate(Routes.WALLET_SCREEN);
-              });
+              }, 300);
               if (android) {
                 hide();
               }
+              analytics.track('Imported seed phrase', {
+                hadPreviousAddressWithValue,
+              });
             } else {
-              handleSetImporting(false);
+              // Wait for error messages then refocus
+              setTimeout(() => {
+                inputRef.current?.focus();
+                initializeWallet();
+              }, 100);
             }
           })
           .catch(error => {
             handleSetImporting(false);
             logger.error('error importing seed phrase: ', error);
+            setTimeout(() => {
+              inputRef.current?.focus();
+              initializeWallet();
+            }, 100);
           });
       }, 50);
     }