+++ src/screens/SendSheet.js
@@ -1,5 +1,6 @@
 import { useRoute } from '@react-navigation/native';
 import analytics from '@segment/analytics-react-native';
+import { captureEvent, captureException } from '@sentry/react-native';
 import { get, isEmpty, isString, toLower } from 'lodash';
 import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { InteractionManager, Keyboard, StatusBar } from 'react-native';
@@ -44,6 +45,7 @@ import {
 import Routes from '@rainbow-me/routes';
 import { borders, colors } from '@rainbow-me/styles';
 import { deviceUtils, gasUtils } from '@rainbow-me/utils';
+import logger from 'logger';
 
 const sheetHeight = deviceUtils.dimensions.height - 10;
 const statusBarHeight = getStatusBarHeight(true);
@@ -237,8 +239,14 @@ export default function SendSheet(props) {
   const onSubmit = useCallback(async () => {
     const validTransaction =
       isValidAddress && amountDetails.isSufficientBalance && isSufficientGas;
-    if (!selectedGasPrice.txFee || !validTransaction || isAuthorizing)
+    if (!selectedGasPrice.txFee || !validTransaction || isAuthorizing) {
+      logger.sentry('preventing tx submit for one of the following reasons:');
+      logger.sentry('selectedGasPrice.txFee ? ', selectedGasPrice?.txFee);
+      logger.sentry('validTransaction ? ', validTransaction);
+      logger.sentry('isAuthorizing ? ', isAuthorizing);
+      captureEvent('Preventing tx submit');
       return false;
+    }
 
     let submitSuccess = false;
 
@@ -262,6 +270,9 @@ export default function SendSheet(props) {
         await dispatch(dataAddNewTransaction(txDetails));
       }
     } catch (error) {
+      logger.sentry('TX Details', txDetails);
+      logger.sentry('SendSheet onSubmit error');
+      captureException(error);
       submitSuccess = false;
     } finally {
       setIsAuthorizing(false);
@@ -283,13 +294,17 @@ export default function SendSheet(props) {
   ]);
 
   const submitTransaction = useCallback(async () => {
-    if (Number(amountDetails.assetAmount) <= 0) return false;
+    if (Number(amountDetails.assetAmount) <= 0) {
+      logger.sentry('amountDetails.assetAmount ? ', amountDetails?.assetAmount);
+      captureEvent('Preventing tx submit due to amount <= 0');
+      return false;
+    }
 
     try {
       const submitSuccessful = await onSubmit();
       analytics.track('Sent transaction', {
-        assetName: selected.name,
-        assetType: selected.type,
+        assetName: selected?.name || '',
+        assetType: selected?.type || '',
         isRecepientENS: toLower(recipient.slice(-4)) === '.eth',
       });
       if (submitSuccessful) {
@@ -303,8 +318,8 @@ export default function SendSheet(props) {
     navigate,
     onSubmit,
     recipient,
-    selected.name,
-    selected.type,
+    selected?.name,
+    selected?.type,
   ]);
 
   const onPressTransactionSpeed = useCallback(