+++ src/utils/logger.js
@@ -1,3 +1,4 @@
+import { captureException } from '@sentry/react-native';
 import sentryUtils from './sentry';
 
 const Logger = {
@@ -39,7 +40,11 @@ const Logger = {
     if (args.length === 1 && typeof args[0] === 'string') {
       sentryUtils.addInfoBreadcrumb.apply(null, args);
     } else {
-      sentryUtils.addDataBreadcrumb.apply(null, args);
+      const safeData = safelyStringifyWithFormat(args[1]);
+      sentryUtils.addDataBreadcrumb(args[0], safeData);
+    }
+    if (__DEV__) {
+      console.log(...args);
     }
   },
   warn(...args) {
@@ -47,6 +52,30 @@ const Logger = {
       console.warn(...args);
     }
   },
+};
+
+const safelyStringifyWithFormat = data => {
+  try {
+    const seen = [];
+    const newData = JSON.stringify(
+      data,
+      // Required to ignore cyclic structures
+      (key, val) => {
+        if (val != null && typeof val == 'object') {
+          if (seen.indexOf(val) >= 0) {
+            return;
+          }
+          seen.push(val);
+        }
+        return val;
+      },
+      2
+    );
+    return { data: newData };
+  } catch (e) {
+    captureException(e);
+    return {};
+  }
 };
 
 export default Logger;