+++ CHANGELOG.md
@@ -7,6 +7,46 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 
 ## [Unreleased]
 ### Added
+* Add biometry icon to backup button
+
+### Changed
+* Move rn-nodeify to dev deps
+* Change dapp names and hide swap button for not swappable tokens
+* Deps clean up
+
+### Removed
+
+## [1.2.27](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.27)
+## Added 
+* Migration v5
+
+## [1.2.26](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.26)
+## Added 
+* Migration v4
+
+## Changed
+* check if hasKey instead of loading it directly
+* Sort experimental keys in Dev Section
+
+## [1.2.25](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.25)
+### Changed
+* Support Wyre order reservations
+* Add migration v3
+
+## [1.2.24](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.24)
+### Added
+* Keychain integrity checks
+* Prevent add funds actions when keychain integrity checks fail
+* Zerion charts integration
+
+### Changed
+* Fix deposit/withdraw modal corners
+* Restore missing icon in send gas button
+* Fix savings deposits
+* Fix qrcode scanner behavior
+
+## [1.2.23](https://github.com/rainbow-me/rainbow/releases/tag/v1.2.23)
+### Added
 * Add welcome screen
 * Add icon on splash screen animated hiding
 * Add tx default data value
@@ -108,6 +148,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/)
 * fix Send flow bugs
 * Fix initial flash happening on Swap output's coin icon placeholder
 * Fix logger import
+* Fix crash when pressing gas speed button
+* Log and restore if possible while showing secret
 
 ### Removed
 * Disable safari debugging